requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain-groq
langchain_community
faiss-cpu
langchain-text-splitters
sentence-transformers
InstructorEmbedding
langchain-huggingface